refactor(providers): move deep-link listener providers into the router layer
auth_providers.dart declared passwordResetListenerProvider and
emailVerificationListenerProvider, so it imported two services that navigate
through goRouterProvider. That put auth_providers — and everything hanging off
it, including environment_provider and moderation_providers — on an import
chain that ran up into the router and back down into the app_providers barrel.

The listeners navigate, so their providers belong in the router layer. Moving
the two declarations to lib/router/providers/deep_link_listeners.dart lets
auth_providers drop both service imports; behaviour and provider identities are
unchanged, and main.dart still initializes both at the same point.

Refs #7383, wave #4506.
